# accounting/openapi/openapi.json
{
...
  "host": "127.0.0.1:45881",
  "info": {
    "title": "Komunitin accounting API",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "description": "API for the Komunitin payments system. This documentation is automatically generated from the serivce test suite."
  },
  "paths": {
...
    "/{code}/topups": {
      "post": {
        "summary": "/{code}/topups",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/topups/{id}": {
      "get": {
        "summary": "/{code}/topups/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "patch": {
        "summary": "/{code}/topups/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/topups/{id}/hooks/mollie": {
      "post": {
        "summary": "/{code}/topups/{id}/hooks/mollie",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/currency/topup-settings": {
      "patch": {
        "summary": "/{code}/currency/topup-settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "currency"
        ]
      },
      "get": {
        "summary": "/{code}/currency/topup-settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "currency"
        ]
      }
    },
    "/{code}/accounts/{id}/topup-settings": {
      "get": {
        "summary": "/{code}/accounts/{id}/topup-settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "summary": "/{code}/accounts/{id}/topup-settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "accounts"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "authorization": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "Bearer JWT from the configured Auth issuer for audience urn:komunitin:api. Use accounting:read for reads, accounting:write for writes, or superadmin on admin operations."
    }
  },
  "servers": [
    {
      "url": "https://accounting.komunitin.org",
      "description": "Production server"
    },
    {
      "url": "https://accounting.demo.komunitin.org",
      "description": "Demo server"
    },
    {
      "url": "http://localhost:2025",
      "description": "Local development server"
    }
  ],
  "swagger": "2.0",
...
}